+++ packages/codemods/src/codemods/v0.39.x/updateBabelConfig/__testfixtures__/defaultConfig/input/web/.babelrc.js
