+++ packages/codemods/src/codemods/v0.39.x/updateBabelConfig/__testfixtures__/webCustom/output/web/babel.config.js
