+++ packages/codemods/src/codemods/v0.39.x/updateBabelConfig/__tests__/updateBabelConfig.test.ts
@@ -1,4 +1,4 @@
-import transform from '../removeBabelConfig'
+import transform from '../updateBabelConfig'
 test('Should remove/rename configs correctly for the default setup', async () => {
   await matchFolderTransform(transform, 'defaultConfig')
 })
