+++ packages/codemods/src/codemods/v0.39.x/updateBabelConfig/updateBabelConfig.ts
