+++ packages/codemods/src/codemods/v0.39.x/updateBabelConfig/updateBabelConfig.yargs.ts
@@ -1,12 +1,12 @@
 import task from 'tasuku'
 
-import { removeBabelConfig } from './removeBabelConfig'
+import { removeBabelConfig } from './updateBabelConfig'
 
-export const command = 'remove-babel-config'
+export const command = 'update-babel-config'
 export const description = '(v0.38-v0.39) Removes babel config from the project'
 
 export const handler = () => {
-  task('Update seed script part', async ({ setOutput }) => {
+  task('Update seed script part', async () => {
     await removeBabelConfig()
 
     // setOutput(notes)
